moved Coinbase and Binance trade classes into their own namespace
