+++ ExchangeSharp/API/Exchanges/_Base/ExchangeAPIExtensions.cs
@@ -14,7 +14,8 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-
+using ExchangeSharp.Binance;
+using ExchangeSharp.Coinbase;
 using Newtonsoft.Json.Linq;
 
 namespace ExchangeSharp
